Cleaned up misc compiler warnings
- added casts where appropriate
- removed deprecated memory management functions that are no-ops

+++ Common/UserLandIACToolkit/landsystem7.c
@@ -571,7 +571,7 @@ static pascal OSErr landsystem7handleevent (AppleEvent *message, AppleEvent *rep
 			
 		#else
 			
-			curA5 = SetUpAppA5 ();
+			curA5 = (long)SetUpAppA5 ();
 		
 		#endif
